backend/auth-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0
httpx==0.27.0
